Fixed InstructorEditClass and ScheduleClass activities
calendars, dates, time is horrible.

+++ app/src/main/java/com/example/finalproject/DB_Management.java
@@ -577,13 +577,13 @@ public class DB_Management extends SQLiteOpenHelper {
                 String s = "";
                 for(int i =1; i < 7; i++){
                     if(i == 3){
-                        Date date = new Date(cursor.getInt(i));
+                        Date date = new Date(cursor.getLong(i));
                         s += dateFormat.format(date);
                         s += " ";
                         s += timeFormat.format(date);
                     }else if(i == 4) {
                         s += "-";
-                        Date date = new Date(cursor.getInt(i));
+                        Date date = new Date(cursor.getLong(i));
                         s += timeFormat.format(date);
                         s += " ";
                     }else {
@@ -621,13 +621,13 @@ public class DB_Management extends SQLiteOpenHelper {
                 String s = "";
                 for(int i =1; i < 7; i++){
                     if(i == 3){
-                        Date date = new Date(cursor.getInt(i));
+                        Date date = new Date(cursor.getLong(i));
                         s += dateFormat.format(date);
                         s += " ";
                         s += timeFormat.format(date);
                     }else if(i == 4) {
                         s += "-";
-                        Date date = new Date(cursor.getInt(i));
+                        Date date = new Date(cursor.getLong(i));
                         s += timeFormat.format(date);
                         s += " ";
                     }else {
@@ -665,13 +665,13 @@ public class DB_Management extends SQLiteOpenHelper {
                 String s = "";
                 for(int i =0; i < 7; i++){
                     if(i == 3){
-                        Date date = new Date(cursor.getInt(i));
+                        Date date = new Date(cursor.getLong(i));
                         s += dateFormat.format(date);
                         s += " ";
                         s += timeFormat.format(date);
                     }else if(i == 4) {
                         s += "-";
-                        Date date = new Date(cursor.getInt(i));
+                        Date date = new Date(cursor.getLong(i));
                         s += timeFormat.format(date);
                         s += " ";
                     }else {
